+++ src/main/java/miyucomics/hexical/casting/spells/OpChorusBlink.kt
@@ -9,8 +9,9 @@ import at.petrak.hexcasting.api.spell.casting.CastingContext
 import at.petrak.hexcasting.api.spell.getVec3
 import at.petrak.hexcasting.api.spell.iota.Iota
 import at.petrak.hexcasting.api.spell.mishaps.MishapLocationTooFarAway
+import miyucomics.hexical.casting.mishaps.NeedsChorusFruitMishap
+import net.minecraft.item.ItemStack
 import net.minecraft.item.Items
-import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.util.math.Vec3d
 
 class OpChorusBlink : SpellAction {
@@ -19,21 +20,20 @@ class OpChorusBlink : SpellAction {
 		val pos = args.getVec3(0, argc)
 		if (pos.lengthSquared() > 256)
 			throw MishapLocationTooFarAway(pos)
-		return Triple(Spell(pos), MediaConstants.DUST_UNIT, listOf())
+		for (stack in DiscoveryHandlers.collectItemSlots(ctx)) {
+			if (stack.isEmpty)
+				continue
+			if (!stack.isOf(Items.CHORUS_FRUIT))
+				continue
+			return Triple(Spell(pos, stack), MediaConstants.DUST_UNIT, listOf())
+		}
+		throw NeedsChorusFruitMishap()
 	}
 
-	private data class Spell(val position: Vec3d) : RenderedSpell {
+	private data class Spell(val position: Vec3d, val stack: ItemStack) : RenderedSpell {
 		override fun cast(ctx: CastingContext) {
-			for (stack in DiscoveryHandlers.collectItemSlots(ctx)) {
-				if (stack.item != Items.CHORUS_FRUIT)
-					continue
-				if (stack.isEmpty)
-					continue
-				stack.decrement(1)
-				val caster: ServerPlayerEntity = ctx.caster
-				caster.teleport(caster.pos.x + position.x, caster.pos.y + position.y, caster.pos.z + position.z)
-				return
-			}
+			stack.decrement(1)
+			ctx.caster.teleport(ctx.caster.pos.x + position.x, ctx.caster.pos.y + position.y, ctx.caster.pos.z + position.z)
 		}
 	}
 }