+++ src/main/java/miyucomics/hexical/items/LightningRodStaff.kt
@@ -9,13 +9,9 @@ import net.minecraft.entity.LivingEntity
 import net.minecraft.entity.attribute.EntityAttribute
 import net.minecraft.entity.attribute.EntityAttributeModifier
 import net.minecraft.entity.attribute.EntityAttributes
-import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.item.ItemStack
 import net.minecraft.sound.SoundCategory
 import net.minecraft.sound.SoundEvents
-import net.minecraft.util.Hand
-import net.minecraft.util.TypedActionResult
-import net.minecraft.world.World
 
 class LightningRodStaff : ItemStaff(Settings().maxCount(1).group(HexicalItems.HEXICAL_GROUP)) {
 	private val attributeModifiers: Multimap<EntityAttribute, EntityAttributeModifier>