+++ src/main/java/miyucomics/hexical/mixin/PlayerEntityMixin.java
@@ -2,9 +2,6 @@ package miyucomics.hexical.mixin;
 
 import miyucomics.hexical.interfaces.PlayerEntityMinterface;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NbtCompound;
-import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;