+++ src/main/java/miyucomics/hexical/state/PersistentStateHandler.kt
@@ -7,31 +7,18 @@ import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.item.ItemStack
 import net.minecraft.nbt.NbtCompound
 import net.minecraft.server.MinecraftServer
-import net.minecraft.util.math.BlockPos
 import net.minecraft.world.PersistentState
 import net.minecraft.world.World
 import java.util.*
 
 class PersistentStateHandler : PersistentState() {
 	private var archLamps: HashMap<UUID, ArchLampData> = HashMap<UUID, ArchLampData>()
-	private var boundLibraries: HashMap<UUID, BlockPos?> = HashMap<UUID, BlockPos?>()
 	private var wristpockets: HashMap<UUID, ItemStack> = HashMap<UUID, ItemStack>()
 
 	override fun writeNbt(nbt: NbtCompound): NbtCompound {
 		val nbtArchLamps = NbtCompound()
 		archLamps.forEach { (uuid, playerData) -> nbtArchLamps.put(uuid.toString(), playerData.toNbt()) }
 		nbt.put("arch_lamp", nbtArchLamps)
-		val nbtBoundLibraries = NbtCompound()
-		boundLibraries.forEach { (uuid, position) ->
-			if (position == null)
-				return@forEach
-			val serializedPosition = NbtCompound()
-			serializedPosition.putInt("x", position.x)
-			serializedPosition.putInt("y", position.y)
-			serializedPosition.putInt("z", position.z)
-			nbtBoundLibraries.putCompound(uuid.toString(), serializedPosition)
-		}
-		nbt.put("bound_libraries", nbtBoundLibraries)
 		val nbtWristpockets = NbtCompound()
 		wristpockets.forEach { (uuid, stack) -> nbtWristpockets.putCompound(uuid.toString(), stack.serializeToNBT()) }
 		nbt.put("wristpockets", nbtWristpockets)
@@ -43,11 +30,6 @@ class PersistentStateHandler : PersistentState() {
 			val state = PersistentStateHandler()
 			val nbtArchLamps = tag.getCompound("arch_lamp")
 			nbtArchLamps.keys.forEach { key: String -> state.archLamps[UUID.fromString(key)] = ArchLampData.createFromNbt(nbtArchLamps.getCompound(key)) }
-			val nbtBoundLibraries = tag.getCompound("bound_libraries")
-			nbtBoundLibraries.keys.forEach { key: String ->
-				val position = nbtBoundLibraries.getCompound(key)
-				state.boundLibraries[UUID.fromString(key)] = BlockPos(position.getInt("x"), position.getInt("y"), position.getInt("z"))
-			}
 			val nbtWristpockets = tag.getCompound("wristpockets")
 			nbtWristpockets.keys.forEach { key: String -> state.wristpockets[UUID.fromString(key)] = ItemStack.fromNbt(nbtWristpockets.getCompound(key)) }
 			return state
@@ -63,16 +45,6 @@ class PersistentStateHandler : PersistentState() {
 		fun getPlayerArchLampData(player: PlayerEntity): ArchLampData {
 			val serverState: PersistentStateHandler = getServerState(player.getWorld().server!!)
 			return serverState.archLamps.computeIfAbsent(player.uuid) { ArchLampData() }
-		}
-
-		fun setPlayerBoundLibrary(player: PlayerEntity, position: BlockPos?) {
-			val serverState: PersistentStateHandler = getServerState(player.getWorld().server!!)
-			serverState.boundLibraries[player.uuid] = position
-		}
-
-		fun getPlayerBoundLibrary(player: PlayerEntity): BlockPos? {
-			val serverState: PersistentStateHandler = getServerState(player.getWorld().server!!)
-			return serverState.boundLibraries.computeIfAbsent(player.uuid) { null }
 		}
 
 		fun stashWristpocket(player: PlayerEntity, stack: ItemStack) {